acpi: clh: Update FACT table to fix guest shutdown
Currently CLH guest were not able to shutdown cleanly because of the
incorrect FACT table. Due to which S5 sleep state was not advertised to
the guest and guest got stuck after invoking the HLT instruction.

Thus, fix the FACP table to correctly advertise the sleep state and
sleep registers.

Signed-off-by: Jinank Jain <jinankjain@microsoft.com>