acpi: clh: Remove multiple ACPI table for CLH guest
On the CVM guest kernel side we have figured out a way to unify multiple
IGVMs for different CPU configuration into one. Thus, it is about time
that we can remove multiple directories that we have under acpi-clh for
different CPU configuration and have a single set of ACPI table for CLH
guests.

Signed-off-by: Jinank Jain <jinankjain@microsoft.com>