igvm: don't hardcode all CPU extended state subleaves
Remove hardcoding of CPU extended state subleaves in the CPUID page in
the IGVM file. Include only the 0x0 and 0x1 subleaves.

For the remaining subleaves, the VMM will query and import only the
supported subleaves.

Hardcoding some unsupported leaves with zeroes can sometimes lead to the
PSP rejecting the CPUID page. One instance we've observed of this is
when AVX512 is supported by the hardware the PSP doesn't allow setting
zeroes for the corresponding subleaves even if AVX512 is not exposed to
the VM. However, the PSP is okay if we just omit those subleaves in our
CPUID page.

Signed-off-by: Anirudh Rayabharam <anrayabh@microsoft.com>